% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_cols_seq.R
\name{impute_cols_seq}
\alias{impute_cols_seq}
\title{Sequentially imputation (columns)}
\usage{
impute_cols_seq(
  ds,
  model_spec_parsnip = linear_reg(),
  cols_order = seq_len(ncol(ds)),
  cols_used_for_imputation = "only_complete",
  rows_order = seq_len(nrow(ds)),
  rows_used_for_imputation = "only_complete",
  M = is.na(ds)
)
}
\arguments{
\item{ds}{the data set to be imputed}

\item{model_spec_parsnip}{The model type used for imputation. It is defined
via the \code{parsnip} package.}

\item{cols_order}{ordering of the columns for imputation}

\item{cols_used_for_imputation}{Which columns should be used to impute other
columns? Possible choices: "only_complete", "already_imputed", "all"}

\item{rows_order}{ordering of the rows for imputation}

\item{rows_used_for_imputation}{Which rows should be used to impute other
rows? Possible choices: "only_complete", "already_imputed", "all"}

\item{M}{missing data indicator matrix}
}
\value{
The imputed data set.
}
\description{
Impute columns of a data set sequentially
}
\details{
This function imputes the columns of the data set \code{ds} column by column. The
imputation order of the column can be specified by \code{cols_order}. Furthermore,
\code{cols_used_for_imputation} controls which columns are used for the
imputation. The same options are available for the rows of \code{ds} via
\code{rows_order} and \code{rows_used_for_imputation}. If \code{ds} is pre-imputed, the
missing data indicator matrix can be supplied via \code{M}.
}
\examples{
# to be done
}
