% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_donor.R
\name{predict_donor}
\alias{predict_donor}
\title{Prediction for donor-based imputation}
\usage{
predict_donor(
  ds_donors,
  ds,
  M = is.na(ds),
  i,
  donor_aggregation = "choose_random"
)
}
\arguments{
\item{ds_donors}{Data set with donors, normally generated by \code{\link[=model_donor]{model_donor()}}}

\item{ds}{The data set to be imputed. Must be a data frame with column names.}

\item{M}{Missing data indicator matrix}

\item{i}{Index of row of \code{ds} which should be imputed}

\item{donor_aggregation}{Type of donor aggregation. Can be one of
'choose_random' and 'average'.}
}
\value{
The imputation values for row \code{i}.
}
\description{
This function is intended to be used inside of \code{\link[=impute_unsupervised]{impute_unsupervised()}} as \code{predict_fun}.
}
\seealso{
\code{\link[=model_donor]{model_donor()}}
}
