% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_fun.R
\name{stop_ds_difference}
\alias{stop_ds_difference}
\title{Compare difference between two data sets}
\usage{
stop_ds_difference(
  ds,
  ds_old,
  info_list,
  stop_eps = 1e-06,
  stop_p = 1,
  stop_sum_diffs = TRUE,
  stop_na_rm = TRUE
)
}
\arguments{
\item{ds}{a numeric data set}

\item{ds_old}{a numeric data set}

\item{info_list}{info_list used inside of \code{\link[=impute_iterative]{impute_iterative()}}. Only the list
element \code{nr_iterations} is used/needed.}

\item{stop_eps}{Threshold value for the difference.}

\item{stop_p}{exponent used for the calculation of differences similar to
Minkowski distance. For \code{stop_p = 1} the absolute differences are used. For
\code{stop_p = 2} the quadratic differences are summed and the square root of
this sum is compared with \code{stop_eps}.}

\item{stop_sum_diffs}{Should differences be summed or averaged
(\code{stop_sum_diffs = FALSE})?}

\item{stop_na_rm}{Should \code{NA}-values be removed when calculating the sum/average?}
}
\value{
\code{FALSE}, if difference is too big. Otherwise \code{ds} with number of
iterations (\code{nr_iterations}) as attribute.
}
\description{
This function is intended to be used as \code{stop_fun} inside of
\code{\link[=impute_iterative]{impute_iterative()}}. It compares the difference of two (numeric) data sets
and return \code{ds}, if difference is small enough (less then \code{stop_es}).
}
\examples{
# See impute_iterative()
}
