% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_cols_seq.R
\name{impute_cols_seq}
\alias{impute_cols_seq}
\title{Sequentially imputation (columns)}
\usage{
impute_cols_seq(
  ds,
  model_spec_parsnip = linear_reg(),
  cols_used_for_imputation = "only_complete",
  cols_order = seq_len(ncol(ds)),
  rows_used_for_imputation = "only_complete",
  rows_order = seq_len(nrow(ds)),
  M = is.na(ds),
  ...
)
}
\arguments{
\item{ds}{the data set to be imputed, must be a data frame with column names}

\item{model_spec_parsnip}{The model type used for imputation. It is defined
via the \code{parsnip} package.}

\item{cols_used_for_imputation}{Which columns should be used to impute other
columns? Possible choices: "only_complete", "already_imputed", "all", "all_no_update"}

\item{cols_order}{ordering of the columns for imputation. This can be a vector with
indices or an \code{order_option} from \code{\link[=order_cols]{order_cols()}}.}

\item{rows_used_for_imputation}{Which rows should be used to impute other
rows? Possible choices: "only_complete", "partly_complete",
"already_imputed", "all_except_i", "all_except_i_no_update", "all",
"all_no_update"}

\item{rows_order}{Ordering of the rows for imputation. This can be a vector with
indices or an \code{order_option} from \code{\link[=order_rows]{order_rows()}}.}

\item{M}{missing data indicator matrix}

\item{...}{arguments passed on to \code{\link[parsnip:reexports]{parsnip::fit_xy()}} and \code{\link[stats:predict]{stats::predict()}}}
}
\value{
The imputed data set.
}
\description{
Impute columns of a data set sequentially
}
\details{
This function imputes the columns of the data set \code{ds} column by
column. The imputation order of the column can be specified by \code{cols_order}.
Furthermore, \code{cols_used_for_imputation} controls which columns are used for
the imputation. The same options are available for the rows of \code{ds} via
\code{rows_order} and \code{rows_used_for_imputation}. If \code{ds} is pre-imputed, the
missing data indicator matrix can be supplied via \code{M}.

The options "all" and "all_no_update" for \code{cols_used_for_imputation} and
"all_except_i", "all_except_i_no_update", "all", "all_no_update" for
\code{rows_used_for_imputation} are only available, if \code{ds} is complete.
If a "_no_update" option is chosen, it will override the other options.
Therefore, the results of e.g. \code{cols_used_for_imputation} = "all_no_update"
combined with \code{rows_used_for_imputation} = "all" or
\code{rows_used_for_imputation} = "all_no_update" are the same.
}
\examples{
# to be done
}
