% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_donor.R
\name{model_donor}
\alias{model_donor}
\title{Model for donor-based imputation}
\usage{
model_donor(ds, M, i, donor_selection = "simultan_complete", donor_k = 10)
}
\arguments{
\item{ds}{the data set to be imputed, must be a data frame with column names}

\item{M}{missing data indicator matrix}

\item{i}{row of i}

\item{donor_selection}{How to select the donors?}

\item{donor_k}{number of selected closest donor}
}
\value{
A "model" for \code{\link[=predict_donor]{predict_donor()}} which is merely a data frame.
}
\description{
This function is intended to be used inside of \code{\link[=impute_cols_sim]{impute_cols_sim()}} as \code{model_fun}.
}
