% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_donor.R
\name{predict_donor}
\alias{predict_donor}
\title{Prediction for donor-based imputation}
\usage{
predict_donor(ds_donors, ds, M, i, donor_aggregation = "choose_random")
}
\arguments{
\item{ds_donors}{data set with donors, normally generated by \code{\link[=model_donor]{model_donor()}}}

\item{ds}{the data set to be imputed, must be a data frame with column names}

\item{M}{missing data indicator matrix}

\item{i}{index of row of \code{ds} which should be imputed}

\item{donor_aggregation}{type of donor aggregation. Can be one of
'choose_random' and 'average'}
}
\value{
The imputation values for row \code{i}
}
\description{
This function is intended to be used inside of \code{\link[=impute_unsupervised]{impute_unsupervised()}} as \code{predict_fun}.
}
