% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_fun.R
\name{stop_ds_difference}
\alias{stop_ds_difference}
\title{Compare differences between two data sets}
\usage{
stop_ds_difference(
  ds,
  ds_old,
  info_list,
  stop_args = list(eps = 1e-06, p = 1, sum_diffs = TRUE, na_rm = TRUE)
)
}
\arguments{
\item{ds}{a numeric data set}

\item{ds_old}{a numeric data set}

\item{info_list}{info_list used inside of \code{\link[=impute_iterative]{impute_iterative()}}. Only the list
element \code{nr_iterations} is used/needed.}

\item{stop_args}{A list with following named components:
\itemize{
\item \code{eps} Threshold value for the difference.
\item \code{p} exponent used for the calculation of differences similar to
Minkowski distance. For \code{p = 1} the absolute differences are used. For
\code{p = 2} the quadratic differences are summed and the square root of
this sum is compared with \code{stop_eps}.
\item \code{sum_diffs} Should differences be summed or averaged (\code{sum_diffs = FALSE})?
\item \code{na_rm} Should \code{NA}-values be removed when calculating the sum/average?
If \code{na_rm = FALSE} and there are \code{NA}s, the function returns \verb{FALSE``.}
}}
}
\value{
\code{FALSE}, if difference is too big. Otherwise \code{ds} with number of
iterations (\code{nr_iterations}) as attribute.
}
\description{
This function is intended to be used as \code{stop_fun} inside of
\code{\link[=impute_iterative]{impute_iterative()}}. It compares the difference of two (numeric) data sets
and return \code{ds}, if difference is small enough (less than \code{stop_args$eps}).
}
\examples{
# See impute_iterative()
}
